chore: refactor WrapperViewController
